refactor(examples): reorganize hello_triangle to hello_cube with OOP structure
- Rename hello_triangle to hello_cube to accurately reflect cube rendering
- Restructure example using OOP approach with App class instead of procedural code
- Split functionality into App class with proper initialization, run, and shutdown methods
- Move main logic from main.cpp to app.cpp for better code organization
- Add proper resource management and cleanup in App destructor
- Improve code maintainability and readability with class-based architecture

The example has been reorganized from a procedural approach to an object-oriented
design, making it easier to maintain and extend. The App class encapsulates all
graphics resources and provides a clean interface for initialization, rendering,
and cleanup. The name change from hello_triangle to hello_cube better reflects
the actual geometry being rendered.
