+++ exemples/hello_cube/shaders/CMakeLists.txt
@@ -1,0 +1,42 @@
+cmake_minimum_required(VERSION 3.25)
+
+set(APP_NAME hello_cube)
+
+find_program(GLSLC glslc HINTS "$ENV{VULKAN_SDK}/bin" "/usr/bin")
+if(NOT GLSLC)
+    find_program(GLSLC glslangValidator HINTS "/usr/bin")
+    if(NOT GLSLC)
+        message(FATAL_ERROR "Could not find glslc or glslangValidator")
+    endif()
+endif()
+
+set(SHADER_DIR "${CMAKE_CURRENT_BINARY_DIR}/shaders")
+file(MAKE_DIRECTORY ${SHADER_DIR})
+
+set(SHADERS cube.vert cube.frag)
+set(SPIRV_FILES)
+
+foreach(SHADER ${SHADERS})
+    set(INPUT "${CMAKE_CURRENT_SOURCE_DIR}/${SHADER}")
+    set(OUTPUT "${SHADER_DIR}/${SHADER}.spv")
+    
+    add_custom_command(
+        OUTPUT ${OUTPUT}
+        COMMAND ${GLSLC} --target-env=vulkan1.4 ${INPUT} -o ${OUTPUT}
+        DEPENDS ${INPUT}
+        COMMENT "Compiling ${SHADER}"
+    )
+    
+    list(APPEND SPIRV_FILES ${OUTPUT})
+endforeach()
+
+add_custom_target(${APP_NAME}_shaders ALL DEPENDS ${SPIRV_FILES})
+
+add_custom_command(
+    TARGET ${APP_NAME}_shaders POST_BUILD
+    COMMAND ${CMAKE_COMMAND} -E copy_directory ${SHADER_DIR} $<TARGET_FILE_DIR:${APP_NAME}>/shaders
+    COMMENT "Copying shaders to executable directory"
+)
+
+message(STATUS "Shaders will be compiled to: ${SHADER_DIR}")
+message(STATUS "and copied to: $<TARGET_FILE_DIR:${APP_NAME}>/shaders")
