--- exemples/hello_triangle/main.cpp
@@ -1,569 +1,0 @@
-#include "vostok/core/logger/logger.hpp"
-#include "vostok/graphics/buffers/texture_loader.hpp"
-#include "vostok/graphics/camera/perspective_camera.hpp"
-#include "vostok/graphics/gpu.hpp"
-#include "vostok/graphics/pipeline.hpp"
-#include "vostok/math/types.hpp"
-#include "vostok/window/window.hpp"
-#include "vostok/window/window_config.hpp"
-
-#include <chrono>
-#include <filesystem>
-#include <iostream>
-#include <memory>
-#include <thread>
-#include <vector>
-
-#ifdef _WIN32
-#include <windows.h>
-#else
-#include <limits.h>
-#include <unistd.h>
-#endif
-
-namespace fs = std::filesystem;
-using namespace vostok;
-
-struct CameraUBO
-{
-    alignas(16) math::Mat4 view;
-    alignas(16) math::Mat4 proj;
-    alignas(16) math::Vec3 position;
-    alignas(4) f32 time;
-};
-
-struct Vertex
-{
-    math::Vec3 position;
-    math::Vec2 uv;
-};
-
-struct HelloTriangleApp
-{
-    Window window;
-    graphics::GPU gpu;
-
-    graphics::VBO<Vertex> vertexBuffer;
-    graphics::IBO<u32> indexBuffer;
-    graphics::Pipeline pipeline;
-
-    graphics::UBO<CameraUBO> cameraUBO;
-    graphics::Texture texture;
-    graphics::PerspectiveCamera camera;
-};
-
-auto getExecutablePath() -> fs::path
-{
-#ifdef _WIN32
-    std::array<wchar_t, MAX_PATH> path{};
-    GetModuleFileNameW(nullptr, static_cast<LPWSTR>(path.data()), MAX_PATH);
-    return fs::path(path.data()).parent_path();
-#else
-    char result[PATH_MAX];
-    ssize_t count = readlink("/proc/self/exe", result, PATH_MAX);
-    return fs::path(std::string(result, (count > 0) ? count : 0)).parent_path();
-#endif
-}
-
-auto findResourcePath(
-    const fs::path &resourceType,
-    const fs::path &resourceName
-) -> fs::path
-{
-    fs::path basePath = getExecutablePath();
-    Logger::debug("Executable directory: {}", basePath.string());
-
-    std::vector<fs::path> searchPaths = {
-        basePath / resourceType / resourceName,
-        basePath / "resources" / resourceType / resourceName,
-        basePath.parent_path() / resourceType / resourceName,
-        basePath.parent_path() / "resources" / resourceType / resourceName,
-        basePath.parent_path() / "bin" / resourceType / resourceName,
-        fs::current_path() / resourceType / resourceName,
-        fs::current_path() / "resources" / resourceType / resourceName,
-        fs::current_path() / "bin" / resourceType / resourceName
-    };
-
-    for (const auto &path : searchPaths) {
-        if (fs::exists(path)) {
-            Logger::debug("Found resource at: {}", path.string());
-            return path;
-        }
-        Logger::trace("Resource not found at: {}", path.string());
-    }
-    Logger::warning(
-        "Resource not found: {}/{}",
-        resourceType.string(),
-        resourceName.string()
-    );
-    return basePath / resourceType / resourceName;
-}
-
-auto initializeLogger() -> bool
-{
-    fs::path executablePath = getExecutablePath();
-    fs::path logDir = executablePath / "logs";
-
-    try {
-        if (!fs::exists(logDir)) {
-            fs::create_directories(logDir);
-        }
-    } catch (const std::exception &e) {
-        std::cerr << "Failed to create log directory: " << e.what() << "\n";
-        logDir = fs::current_path() / "logs";
-        try {
-            if (!fs::exists(logDir)) {
-                fs::create_directories(logDir);
-            }
-        } catch (const std::exception &e) {
-            std::cerr << "Failed to create log directory in working path: "
-                      << e.what() << "\n";
-            logDir = fs::current_path();
-        }
-    }
-    LogConfig logConfig;
-    logConfig.asyncMode = false;
-    logConfig.level = LogLevel::DEBUG;
-    logConfig.file = FileLogConfig();
-    logConfig.file->filePath = (logDir / "vostok_default.log").string();
-    logConfig.file->truncateOnStart = true;
-    logConfig.file->separateFilesByComponent = true;
-    logConfig.file->componentFilePattern =
-        (logDir / "vostok_{name}.log").string();
-
-    auto result = Logger::init(logConfig);
-    if (!result) {
-        std::cerr << "Failed to initialize logger: " << result.error() << "\n";
-        return false;
-    }
-
-    Logger::info("Logger initialized");
-    Logger::info("Logs will be stored in: {}", logDir.string());
-    return true;
-}
-
-auto createWindow() -> std::expected<Window, std::string>
-{
-    WindowConfig windowInfo;
-    windowInfo.title = "Vostok Hello Triangle";
-    windowInfo.width = 800;
-    windowInfo.height = 600;
-    windowInfo.resizable = true;
-
-    return Window::create(windowInfo);
-}
-
-auto createGPUDevice(WindowHandle *window)
-    -> std::expected<graphics::GPU, std::string>
-{
-    graphics::GPUHandle::CreateInfo deviceInfo;
-    deviceInfo.appName = "Vostok Hello Triangle";
-    deviceInfo.appVersion = core::Version{ .major = 0, .minor = 1, .patch = 0 };
-    deviceInfo.enableValidationLayers = true;
-    deviceInfo.windowHandle = window->getNativeHandle();
-    deviceInfo.width = window->getWidth();
-    deviceInfo.height = window->getHeight();
-
-    return graphics::GPU::create(deviceInfo);
-}
-
-auto createVertexBuffer(graphics::GPUHandle *gpu)
-    -> std::expected<graphics::VBO<Vertex>, std::string>
-{
-    std::vector<Vertex> vertices = {
-        // Front face (z = 0.5)
-        { .position = { -0.5F, -0.5F, 0.5F }, .uv = { 0.0F, 0.0F } },
-        { .position = { 0.5F, -0.5F, 0.5F }, .uv = { 1.0F, 0.0F } },
-        { .position = { 0.5F, 0.5F, 0.5F }, .uv = { 1.0F, 1.0F } },
-        { .position = { -0.5F, 0.5F, 0.5F }, .uv = { 0.0F, 1.0F } },
-
-        // Back face (z = -0.5)
-        { .position = { 0.5F, -0.5F, -0.5F }, .uv = { 0.0F, 0.0F } },
-        { .position = { -0.5F, -0.5F, -0.5F }, .uv = { 1.0F, 0.0F } },
-        { .position = { -0.5F, 0.5F, -0.5F }, .uv = { 1.0F, 1.0F } },
-        { .position = { 0.5F, 0.5F, -0.5F }, .uv = { 0.0F, 1.0F } },
-
-        // Left face (x = -0.5)
-        { .position = { -0.5F, -0.5F, -0.5F }, .uv = { 0.0F, 0.0F } },
-        { .position = { -0.5F, -0.5F, 0.5F }, .uv = { 1.0F, 0.0F } },
-        { .position = { -0.5F, 0.5F, 0.5F }, .uv = { 1.0F, 1.0F } },
-        { .position = { -0.5F, 0.5F, -0.5F }, .uv = { 0.0F, 1.0F } },
-
-        // Right face (x = 0.5)
-        { .position = { 0.5F, -0.5F, 0.5F }, .uv = { 0.0F, 0.0F } },
-        { .position = { 0.5F, -0.5F, -0.5F }, .uv = { 1.0F, 0.0F } },
-        { .position = { 0.5F, 0.5F, -0.5F }, .uv = { 1.0F, 1.0F } },
-        { .position = { 0.5F, 0.5F, 0.5F }, .uv = { 0.0F, 1.0F } },
-
-        // Bottom face (y = -0.5)
-        { .position = { -0.5F, -0.5F, -0.5F }, .uv = { 0.0F, 0.0F } },
-        { .position = { 0.5F, -0.5F, -0.5F }, .uv = { 1.0F, 0.0F } },
-        { .position = { 0.5F, -0.5F, 0.5F }, .uv = { 1.0F, 1.0F } },
-        { .position = { -0.5F, -0.5F, 0.5F }, .uv = { 0.0F, 1.0F } },
-
-        // Top face (y = 0.5)
-        { .position = { -0.5F, 0.5F, 0.5F }, .uv = { 0.0F, 0.0F } },
-        { .position = { 0.5F, 0.5F, 0.5F }, .uv = { 1.0F, 0.0F } },
-        { .position = { 0.5F, 0.5F, -0.5F }, .uv = { 1.0F, 1.0F } },
-        { .position = { -0.5F, 0.5F, -0.5F }, .uv = { 0.0F, 1.0F } }
-    };
-
-    return gpu->createVBO<Vertex>(
-        vertices,
-        graphics::formats::VEC3,
-        graphics::formats::VEC2
-    );
-}
-
-auto createIndexBuffer(graphics::GPUHandle *gpu)
-    -> std::expected<graphics::IBO<u32>, std::string>
-{
-    std::vector<u32> indices = { // Front face (z = 0.5)
-                                 0,
-                                 1,
-                                 2,
-                                 2,
-                                 3,
-                                 0,
-                                 // Back face (z = -0.5)
-                                 4,
-                                 5,
-                                 6,
-                                 6,
-                                 7,
-                                 4,
-                                 // Left face (x = -0.5)
-                                 8,
-                                 9,
-                                 10,
-                                 10,
-                                 11,
-                                 8,
-                                 // Right face (x = 0.5)
-                                 12,
-                                 13,
-                                 14,
-                                 14,
-                                 15,
-                                 12,
-                                 // Bottom face (y = -0.5)
-                                 16,
-                                 17,
-                                 18,
-                                 18,
-                                 19,
-                                 16,
-                                 // Top face (y = 0.5)
-                                 20,
-                                 21,
-                                 22,
-                                 22,
-                                 23,
-                                 20
-    };
-
-    return gpu->createIBO<u32>(indices);
-}
-
-auto createTestTexture(graphics::GPUHandle *gpu)
-    -> std::expected<graphics::Texture, std::string>
-{
-    return graphics::TextureLoader::loadFromFile(
-        gpu,
-        findResourcePath("textures", "test.png")
-    );
-}
-
-auto createPipeline(
-    graphics::GPUHandle *gpu,
-    graphics::VBOImpl<Vertex> *vertexBuffer
-) -> std::expected<graphics::Pipeline, std::string>
-{
-    fs::path vertexShaderPath =
-        findResourcePath("shaders", "triangle.vert.spv");
-    fs::path fragmentShaderPath =
-        findResourcePath("shaders", "triangle.frag.spv");
-
-    if (!fs::exists(vertexShaderPath) || !fs::exists(fragmentShaderPath)) {
-        std::vector<fs::path> shaderLocations = {
-            "shaders",
-            "bin/shaders",
-            "../bin/shaders",
-            "build/bin/shaders",
-            "../../../build/bin/shaders"
-        };
-
-        bool found = false;
-        for (const auto &dir : shaderLocations) {
-            auto vert = dir / "triangle.vert.spv";
-            auto frag = dir / "triangle.frag.spv";
-
-            if (fs::exists(vert) && fs::exists(frag)) {
-                vertexShaderPath = vert;
-                fragmentShaderPath = frag;
-                found = true;
-                Logger::info(
-                    "Found shaders in fallback location: {}",
-                    dir.string()
-                );
-                break;
-            }
-        }
-
-        if (!found) {
-            return std::unexpected("Could not find shader files");
-        }
-    } else {
-        Logger::info("Found shaders:");
-        Logger::info("  Vertex shader: {}", vertexShaderPath.string());
-        Logger::info("  Fragment shader: {}", fragmentShaderPath.string());
-    }
-
-    graphics::Pipeline::CreateInfo pipelineInfo;
-    pipelineInfo.name = "TrianglePipeline";
-    pipelineInfo.vertexShader = vertexShaderPath;
-    pipelineInfo.fragmentShader = fragmentShaderPath;
-    pipelineInfo.primitiveTopology = graphics::PrimitiveTopology::TRIANGLE_LIST;
-    pipelineInfo.polygonMode = graphics::PolygonMode::FILL;
-    pipelineInfo.cullMode = graphics::CullMode::NONE;
-    pipelineInfo.frontFace = graphics::FrontFace::COUNTER_CLOCKWISE;
-    pipelineInfo.lineWidth = 1.0F;
-
-    pipelineInfo.depthTest = true;
-    pipelineInfo.depthWrite = true;
-    pipelineInfo.depthCompareOp = graphics::CompareOp::LESS;
-    pipelineInfo.stencilTest = false;
-
-    pipelineInfo.depthFormat = graphics::ImageFormat::D32_SFLOAT;
-
-    pipelineInfo.blend = true;
-    pipelineInfo.srcColorBlendFactor = graphics::BlendFactor::SRC_ALPHA;
-    pipelineInfo.dstColorBlendFactor =
-        graphics::BlendFactor::ONE_MINUS_SRC_ALPHA;
-    pipelineInfo.srcAlphaBlendFactor = graphics::BlendFactor::ONE;
-    pipelineInfo.dstAlphaBlendFactor = graphics::BlendFactor::ZERO;
-    pipelineInfo.colorBlendOp = graphics::BlendOp::ADD;
-    pipelineInfo.alphaBlendOp = graphics::BlendOp::ADD;
-    pipelineInfo.colorWriteMask = graphics::ColorComponentFlags::ALL;
-
-    pipelineInfo.pushConstantSize = sizeof(math::Vec3);
-    pipelineInfo.vertexLayout = vertexBuffer->getLayout();
-
-    return gpu->createPipeline(pipelineInfo);
-}
-
-auto createUBO(graphics::GPUHandle *gpu) -> std::expected<
-    std::pair<graphics::UBO<CameraUBO>, graphics::PerspectiveCamera>,
-    std::string>
-{
-    graphics::PerspectiveCamera::CreateInfo cameraInfo;
-    cameraInfo.name = "MainCamera";
-    cameraInfo.position = { 0.0F, 0.0F, 3.0F };
-    cameraInfo.rotation = { 1.0F, 0.0F, 0.0F, 0.0F };
-    cameraInfo.perspective.fieldOfView = 45.0F;
-    cameraInfo.perspective.aspectRatio = 800.0F / 600.0F;
-    cameraInfo.perspective.nearPlane = 0.1F;
-    cameraInfo.perspective.farPlane = 100.0F;
-
-    auto camera = graphics::PerspectiveCamera::create(cameraInfo);
-
-    CameraUBO initialData{};
-    initialData.view = camera->getViewMatrix();
-    initialData.proj = camera->getProjectionMatrix();
-    initialData.position = camera->getPosition();
-    initialData.time = 0.0F;
-
-    auto uboResult = gpu->createUBO<CameraUBO>(initialData);
-    if (!uboResult) {
-        return std::unexpected(uboResult.error());
-    }
-
-    return std::make_pair(std::move(uboResult.value()), std::move(camera));
-}
-
-auto mainLoop(HelloTriangleApp &app) -> void
-{
-    Logger::info("Entering main render loop");
-
-    uint32_t frameCount = 0;
-    auto startTime = std::chrono::high_resolution_clock::now();
-
-    while (!app.window->shouldClose()) {
-        try {
-            app.window->pollEvents();
-
-            auto frameResult = app.gpu->beginFrame();
-            if (frameResult) {
-                app.pipeline->bind();
-
-                const f32 T = static_cast<f32>(frameCount) * 0.02F;
-                if (app.camera && app.cameraUBO) {
-                    const f32 RADIUS = 3.0F;
-                    const f32 HEIGHT = 1.0F;
-
-                    app.camera->setPosition(
-                        { RADIUS * std::cos(T),
-                          HEIGHT + (0.5F * std::sin(T * 2.0F)),
-                          RADIUS * std::sin(T) }
-                    );
-
-                    app.camera->lookAt({ .target = { 0.0F, 0.0F, 0.0F } });
-
-                    app.cameraUBO->view = app.camera->getViewMatrix();
-                    app.cameraUBO->proj = app.camera->getProjectionMatrix();
-                    app.cameraUBO->position = app.camera->getPosition();
-                    app.cameraUBO->time = T;
-                }
-
-                app.vertexBuffer->bind();
-                app.indexBuffer->bind();
-
-                app.gpu->drawIndexed(app.indexBuffer->getIndexCount());
-
-                auto endResult = app.gpu->endFrame();
-                if (!endResult) {
-                    Logger::warning(
-                        "Failed to end frame: {}",
-                        endResult.error()
-                    );
-                }
-
-                frameCount++;
-            } else {
-                Logger::warning(
-                    "Failed to begin frame: {}",
-                    frameResult.error()
-                );
-
-                if (frameResult.error().find("out of date") !=
-                        std::string::npos ||
-                    frameResult.error().find("Surface lost") !=
-                        std::string::npos) {
-                    Logger::info(
-                        "Swapchain needs recreation, waiting for window "
-                        "resize..."
-                    );
-
-                    std::this_thread::sleep_for(std::chrono::milliseconds(100));
-                } else {
-                    std::this_thread::sleep_for(std::chrono::milliseconds(1));
-                }
-            }
-
-            std::this_thread::sleep_for(std::chrono::milliseconds(16));
-        } catch (const std::exception &e) {
-            Logger::critical("Exception in main loop: {}", e.what());
-            break;
-        } catch (...) {
-            Logger::critical("Unknown exception in main loop");
-            break;
-        }
-    }
-
-    try {
-        app.gpu->waitIdle();
-        Logger::info("Exiting main render loop");
-    } catch (const std::exception &e) {
-        Logger::critical("Exception during GPU wait: {}", e.what());
-    }
-}
-
-auto main(int argc, char *argv[]) -> int
-{
-    if (!initializeLogger()) {
-        return -1;
-    }
-
-    Logger::info("Vostok Hello Triangle starting");
-    Logger::info("Working directory: {}", fs::current_path().string());
-
-    try {
-        HelloTriangleApp app;
-
-        auto windowResult = createWindow();
-        if (!windowResult) {
-            Logger::error("Failed to create window: {}", windowResult.error());
-            return -1;
-        }
-        app.window = std::move(windowResult.value());
-        Logger::info(
-            "Window created successfully: {}x{}",
-            app.window->getWidth(),
-            app.window->getHeight()
-        );
-
-        auto gpuResult = createGPUDevice(app.window.get());
-        if (!gpuResult) {
-            Logger::error("Failed to create GPU device: {}", gpuResult.error());
-            return -1;
-        }
-        app.gpu = std::move(gpuResult.value());
-        Logger::info("GPU device created successfully");
-
-        auto vertexBufferResult = createVertexBuffer(app.gpu.get());
-        if (!vertexBufferResult) {
-            Logger::error(
-                "Failed to create vertex buffer: {}",
-                vertexBufferResult.error()
-            );
-            return -1;
-        }
-        app.vertexBuffer = std::move(vertexBufferResult.value());
-        Logger::info("Vertex buffer created successfully");
-
-        auto indexBufferResult = createIndexBuffer(app.gpu.get());
-        if (!indexBufferResult) {
-            Logger::error(
-                "Failed to create index buffer: {}",
-                indexBufferResult.error()
-            );
-            return -1;
-        }
-        app.indexBuffer = std::move(indexBufferResult.value());
-        Logger::info("Index buffer created successfully");
-
-        auto textureResult = createTestTexture(app.gpu.get());
-        if (!textureResult) {
-            Logger::error(
-                "Failed to create texture: {}",
-                textureResult.error()
-            );
-            return -1;
-        }
-        app.texture = std::move(textureResult.value());
-        Logger::info("Texture created successfully");
-
-        auto pipelineResult =
-            createPipeline(app.gpu.get(), app.vertexBuffer.get());
-
-        if (!pipelineResult) {
-            Logger::error(
-                "Failed to create pipeline: {}",
-                pipelineResult.error()
-            );
-            return -1;
-        }
-        app.pipeline = std::move(pipelineResult.value());
-        Logger::info("Pipeline created successfully");
-
-        auto uboResult = createUBO(app.gpu.get());
-        if (!uboResult) {
-            Logger::error("Failed to create UBO: {}", uboResult.error());
-            return -1;
-        }
-        app.cameraUBO = std::move(uboResult.value().first);
-        app.camera = std::move(uboResult.value().second);
-        Logger::info("UBO created successfully");
-
-        mainLoop(app);
-
-        Logger::info("Application exited successfully");
-    } catch (const std::exception &e) {
-        Logger::critical("Unhandled exception: {}", e.what());
-        return -1;
-    }
-
-    Logger::info("Shutting down logger");
-    Logger::shutdown();
-    return 0;
-}
